rename file direct line tracer
